rag_agent/requirements.txt
fastapi
uvicorn
faiss-cpu 
sentence-transformers
numpy
pydantic
google-generativeai
langchain
langchain_community
